dailybuild: ビルドパラメータ（ライブラリ無視指定）の変更、及び軽微なwarningの抑止
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7127 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/Physics/Joints/GearHandler.h
@@ -40,8 +40,8 @@ public:
 		g.solid->AddShape(phSdk->CreateShape(md));
 		
 		CDCapsuleDesc cd;
-		cd.radius = 0.1;
-		cd.length = l;
+		cd.radius = 0.1f;
+		cd.length = (float)l;
 		g.solid->AddShape(phSdk->CreateShape(cd));
 		g.solid->SetShapePose(1, Posed(Vec3d(r-0.1, 0.0, 0.0), Quaterniond()));
